requirements.txt
streamlit
langchain
langchain-openai
langchain-community
pandas
python-dotenv
typing-extensions
nltk